backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.20.3
pandas==1.3.3
scikit-learn==0.24.2
surprise==1.1.1
matplotlib==3.4.3
seaborn==0.11.2
requests==2.26.0
h5py==3.1.0
tqdm==4.62.3
spotipy==2.19.0
anthropic==0.3.6
python-dotenv==0.19.1 